+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 - New `try` output, which is a drop-in replacement for a `broker` with the `try`
   pattern.
+- Field `successful_on` added to the `http` processor.
 
 ### Changed